requirements.txt
numpy
matplotlib
pandas
jupyter
notebook
pytorch
torchvision